Remove pytest requirement from commands
The link_module and get_outputs functions were using the Proxy class which was in a module that required pytest. This splits things up so that pytest is no longer needed unless you actually want to use pytest-related things.
