+++ pretf/pretf/test.py
@@ -3,172 +3,12 @@ import contextlib
 import functools
 import inspect
 import os
-import sys
 from json import dump as json_dump
-from json import loads as json_loads
-from pathlib import Path
-from subprocess import CompletedProcess
-from types import TracebackType
-from typing import Any, Callable, Dict, Generator, List, Optional, Type, Union
+from typing import Any, Callable, Dict, Generator, List, Type
 
 import pytest
 
-from pretf import render, util, workflow
-
-
-class SensitiveValue:
-    def __init__(self, value: Any):
-        self.value = value
-
-
-class TerraformProxy:
-    def __init__(self, cwd: Union[Path, str] = "", verbose: Optional[bool] = False):
-        if not isinstance(cwd, Path):
-            cwd = Path(cwd)
-        self.cwd = cwd
-        self.env = os.environ.copy()
-        self.env["TF_IN_AUTOMATION"] = "1"
-        self.env["PRETF_VERBOSE"] = "1" if verbose else "0"
-        self.verbose = verbose
-
-    # Calling the object just returns another object with the specified path.
-
-    def __call__(self, cwd: Union[Path, str] = "") -> "TerraformProxy":
-        return self.__class__(cwd or self.cwd)
-
-    # Context manager.
-    # It doesn't do anything but can make the test code easier to follow.
-
-    def __enter__(self) -> "TerraformProxy":
-        return self
-
-    def __exit__(
-        self,
-        exc_type: Optional[Type[BaseException]],
-        exc_val: Optional[BaseException],
-        exc_tb: Optional[TracebackType],
-    ) -> None:
-        return None
-
-    # Terraform command.
-
-    def execute(self, *args: str) -> CompletedProcess:
-        return workflow.execute_terraform(
-            args=args,
-            cwd=self.cwd,
-            env=self.env,
-            capture=True,
-            verbose=self.verbose,
-        )
-
-    # Terraform shortcuts.
-
-    def apply(self, *args: str) -> dict:
-        """
-        Runs terraform apply, parses the output for output values,
-        and returns them as a dictionary.
-
-        """
-
-        apply_args = ["apply", "-json", "-auto-approve=true"]
-        for arg in args:
-            if arg not in apply_args:
-                apply_args.append(arg)
-
-        proc = self.execute(*apply_args)
-
-        outputs = None
-        for line in proc.stdout.splitlines():
-            log = json_loads(line)
-            if log["type"] == "outputs":
-                outputs = log["outputs"]
-
-        if outputs is None:
-            if proc.stderr:
-                print(proc.stderr, file=sys.stderr)
-            raise ValueError(f"Could not parse outputs from: {proc.stdout}")
-
-        values = {}
-
-        for name in outputs:
-            value = outputs[name]["value"]
-            if outputs[name]["sensitive"]:
-                value = SensitiveValue(value)
-            values[name] = value
-
-        return values
-
-    def destroy(self, *args: str) -> str:
-        """
-        Runs terraform destroy and returns the stdout.
-
-        """
-
-        destroy_args = ["destroy", "-input=false", "-auto-approve=true", "-no-color"]
-        for arg in args:
-            if arg not in destroy_args:
-                destroy_args.append(arg)
-        return self.execute(*destroy_args).stdout
-
-    def get(self, *args: str) -> str:
-        """
-        Runs terraform get and returns the stdout.
-
-        """
-
-        get_args = ["get", "-no-color"]
-        for arg in args:
-            if arg not in get_args:
-                get_args.append(arg)
-        return self.execute(*get_args).stdout
-
-    def init(self, *args: str) -> str:
-        """
-        Runs terraform init and returns the stdout.
-
-        """
-
-        init_args = ["init", "-input=false", "-no-color"]
-        for arg in args:
-            if arg not in init_args:
-                init_args.append(arg)
-        return self.execute(*init_args).stdout
-
-    def output(self, *args: str) -> dict:
-        """
-        Runs terraform output and returns the JSON.
-
-        """
-
-        output_args = ["output", "-json"]
-        for arg in args:
-            if arg not in output_args:
-                output_args.append(arg)
-        return json_loads(self.execute(*output_args).stdout)
-
-    def plan(self, *args: str) -> str:
-        """
-        Runs terraform plan and returns the stdout.
-
-        """
-
-        plan_args = ["plan", "-input=false", "-no-color"]
-        for arg in args:
-            if arg not in plan_args:
-                plan_args.append(arg)
-        return self.execute(*plan_args).stdout
-
-
-class PretfProxy(TerraformProxy):
-    def execute(self, *args: str) -> CompletedProcess:
-        return util.execute(
-            file="pretf",
-            args=["pretf"] + list(args),
-            cwd=self.cwd,
-            env=self.env,
-            capture=True,
-            verbose=self.verbose,
-        )
+from pretf import command, render
 
 
 class SimpleTestMeta(type):
@@ -203,6 +43,45 @@ class SimpleTestMeta(type):
         for name, value in list(dct.items()):
             if hasattr(value, "_always"):
                 self._always.add(value.__name__)
+
+
+class SimpleTest(metaclass=SimpleTestMeta):
+
+    pretf = command.PretfCommand()
+    tf = command.TerraformCommand()
+
+    @contextlib.contextmanager
+    def create(self, file_name: str) -> Generator[None, None, None]:
+
+        assert file_name.endswith(".tf.json") or file_name.endswith(".tfvars.json")
+        assert "/" not in file_name
+
+        if not hasattr(self, "_blocks"):
+            self._blocks: Dict[str, list] = collections.defaultdict(list)
+
+        if not hasattr(self, "_create"):
+            self._create: List[str] = []
+
+        self._create.append(file_name)
+
+        yield
+
+        self._create.pop()
+
+        contents = self._blocks.pop(file_name)
+
+        with open(file_name, "w") as open_file:
+            json_dump(contents, open_file, indent=2, default=render.json_default)
+
+
+def always(func: Callable) -> Callable:
+    """
+    Marks a test method to run even when previous tests have failed.
+
+    """
+
+    func._always = True  # type: ignore
+    return func
 
 
 def pretf_test_function(func: Callable) -> Callable:
@@ -255,42 +134,3 @@ def pretf_test_function(func: Callable) -> Callable:
             os.chdir(cwd_before)
 
     return wrapped
-
-
-class SimpleTest(metaclass=SimpleTestMeta):
-
-    pretf = PretfProxy()
-    tf = TerraformProxy()
-
-    @contextlib.contextmanager
-    def create(self, file_name: str) -> Generator[None, None, None]:
-
-        assert file_name.endswith(".tf.json") or file_name.endswith(".tfvars.json")
-        assert "/" not in file_name
-
-        if not hasattr(self, "_blocks"):
-            self._blocks: Dict[str, list] = collections.defaultdict(list)
-
-        if not hasattr(self, "_create"):
-            self._create: List[str] = []
-
-        self._create.append(file_name)
-
-        yield
-
-        self._create.pop()
-
-        contents = self._blocks.pop(file_name)
-
-        with open(file_name, "w") as open_file:
-            json_dump(contents, open_file, indent=2, default=render.json_default)
-
-
-def always(func: Callable) -> Callable:
-    """
-    Marks a test method to run even when previous tests have failed.
-
-    """
-
-    func._always = True  # type: ignore
-    return func
